+++ analysis/analysis_utils.py
@@ -1,5 +1,20 @@
+import asyncio, glob, json, os, time
+
+
+from pathlib import Path
+
+from analysis.kbb import get_pricing_data
+from analysis.normalization import (
+    filter_valid_listings,
+    get_variant_map,
+    normalize_listing,
+)
+from analysis.analysis_utils import check_missing_docs, get_report_dir, get_vehicle_dir
+
+from utils.cache import load_cache
 from utils.common import make_string_url_safe
 from utils.constants import *
+from utils.download import download_files, download_report_pdfs
 from utils.models import TrimValuation
 
 
@@ -117,3 +132,70 @@ def get_trim_valuations_from_cache(
 
             trim_valuations.append(TrimValuation.from_dict(entry))
     return trim_valuations
+
+
+def get_vehicle_dir(listing: dict) -> Path | None:
+    title = listing.get("title")
+    vin = listing.get("vin")
+    if title is None or vin is None:
+        return None
+    path = Path(DOC_PATH) / title / vin
+    return path if path.is_dir() else None
+
+
+def get_report_dir(listing: dict) -> Path | None:
+    dir = get_vehicle_dir(listing)
+    return dir / "carfax.html" if dir else None
+
+
+def check_missing_docs(listings: list[dict]):
+    # Check to see if files exists
+    missing_reports = []
+    for l in listings:
+        dir = get_vehicle_dir(l)
+        if dir is None:
+            continue
+        html = dir / "carfax.html"
+        pdf = dir / "carfax.pdf"
+
+        carfax_url = l.get("additional_docs", {}).get("carfax_url", "Unavailable")
+        if carfax_url != "Unavailable" and not pdf.exists() and not html.exists():
+            missing_reports.append(l)
+
+    if missing_reports:
+        print(f"Downloading reports for {len(missing_reports)} listings...")
+        download_report_pdfs(missing_reports)
+
+
+async def prepare_advanced_analysis(
+    make: str, model: str, listings: list[dict], filename: str
+) -> tuple[list[dict], dict]:
+    cache = load_cache(PRICING_CACHE)
+    cache_entries: dict = cache.setdefault("entries", {})
+
+    # We must normalize listings before getting the variant map
+    norm_listings = [normalize_listing(l) for l in listings]
+    variant_map = await get_variant_map(make, model, norm_listings)
+
+    # Ensure all folders exist, and if not, save the documents
+    if not all(get_vehicle_dir(l) for l in listings):
+        await download_files(listings, filename)
+
+    # Check for missings documents (pdfs, html)
+    check_missing_docs(listings)
+
+    # Filter out only the listings that have a valid report
+    filtered_listings = []
+    for vl in listings:
+        report = get_report_dir(vl)
+        if report and report.exists():
+            filtered_listings.append(normalize_listing(vl))
+
+    # We do not need the trim valuations, we just need to make sure the pricing data has been populated
+    _ = await get_pricing_data(make, model, norm_listings, variant_map, cache)
+
+    valid_listings, _, _ = filter_valid_listings(
+        make, model, filtered_listings, cache_entries, variant_map
+    )
+
+    return (valid_listings, cache_entries)
